+++ include/nmm/frame_state.h
@@ -20,13 +20,13 @@ NMM_API struct nmm_codon_table const* nmm_frame_state_codon_table(
 NMM_API struct nmm_frame_state const* nmm_frame_state_create(char const*                   name,
                                                              struct nmm_base_table const*  baset,
                                                              struct nmm_codon_table const* codont,
-                                                             double                        epsilon);
-NMM_API double                        nmm_frame_state_decode(struct nmm_frame_state const* state,
-                                                             struct imm_seq const* seq, struct nmm_codon* codon);
+                                                             float                         epsilon);
+NMM_API float nmm_frame_state_decode(struct nmm_frame_state const* state, struct imm_seq const* seq,
+                                     struct nmm_codon* codon);
 NMM_API struct nmm_frame_state const* nmm_frame_state_derived(struct imm_state const* state);
 NMM_API void                          nmm_frame_state_destroy(struct nmm_frame_state const* state);
-NMM_API double                        nmm_frame_state_epsilon(struct nmm_frame_state const* state);
-NMM_API double                  nmm_frame_state_lposterior(struct nmm_frame_state const* state,
+NMM_API float                         nmm_frame_state_epsilon(struct nmm_frame_state const* state);
+NMM_API float                   nmm_frame_state_lposterior(struct nmm_frame_state const* state,
                                                            struct nmm_codon const* codon, struct imm_seq const* seq);
 NMM_API struct imm_state const* nmm_frame_state_read(FILE* stream, struct nmm_model const* model);
 NMM_API struct imm_state const* nmm_frame_state_super(struct nmm_frame_state const* state);
