+++ test/src/codon_lprob.c
@@ -23,11 +23,11 @@ void test_codonp(void)
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'C', 'C')) == 0);
 
     cass_cond(imm_lprob_is_zero(nmm_codon_lprob_get(codonp, codon)));
-    cass_cond(nmm_codon_lprob_set(codonp, codon, log(0.5)) == 0);
-    cass_close(nmm_codon_lprob_get(codonp, codon), log(0.5));
+    cass_cond(nmm_codon_lprob_set(codonp, codon, logf(0.5)) == 0);
+    cass_close(nmm_codon_lprob_get(codonp, codon), logf(0.5));
 
     cass_cond(nmm_codon_lprob_normalize(codonp) == 0);
-    cass_close(nmm_codon_lprob_get(codonp, codon), log(1.0));
+    cass_close(nmm_codon_lprob_get(codonp, codon), logf(1.0));
 
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'C', 'X')) == 0);
     cass_cond(!imm_lprob_is_valid(nmm_codon_lprob_get(codonp, codon)));
